split every files to right folder
