+++ src/parts-of-trip-edit-template/format-destination.js
