+++ src/parts-of-trip-edit-template/format-edit-offers.js
