+++ src/parts-of-trip-edit-template/format-pictures.js
