+++ src/parts-of-trip-template/format-offers.js
